Reorganized file layout
