+++ src/components/SampleLayout/SampleLayoutUtils.ts
@@ -46,7 +46,6 @@ export const SampleInitFactoryWebGPU = async (
     const devicePixelRatio = window.devicePixelRatio || 1;
     canvas.width = canvas.clientWidth * devicePixelRatio;
     canvas.height = canvas.clientHeight * devicePixelRatio;
-    console.log(canvas.clientWidth * devicePixelRatio);
     const presentationFormat = navigator.gpu.getPreferredCanvasFormat();
     context.configure({
       device,
