+++ src/pages/samples/[slug].tsx
@@ -12,25 +12,29 @@ type Props = {
 export const pages = {
   simpleVertex: dynamic(() => import('../../sample/simpleVertex/main')),
   normalMapping: dynamic(() => import('../../sample/normalMapping/main')),
-  comanche: dynamic(() => import('../../sample/comancheVoxel/main')),
-  mixExample: dynamic(() => import('../../sample/mixExample/main')),
-  CRT: dynamic(() => import('../../sample/crt/main')),
-  grid: dynamic(() => import('../../sample/grid/main')),
-  cyberpunkGrid: dynamic(() => import('../../sample/cyberpunkGrid/main')),
+  comanche: dynamic(() => import('../../sample/fragment/comancheVoxel/main')),
+  mixExample: dynamic(() => import('../../sample/fragment/mixExample/main')),
+  CRT: dynamic(() => import('../../sample/fragment/crt/main')),
+  grid: dynamic(() => import('../../sample/fragment/grid/main')),
+  cyberpunkGrid: dynamic(
+    () => import('../../sample/fragment/cyberpunkGrid/main')
+  ),
   gltf: dynamic(() => import('../../sample/gltfViewer/main')),
-  computeBalls: dynamic(() => import('../../sample/computeBalls/main')),
+  computeBalls: dynamic(() => import('../../sample/compute/computeBalls/main')),
   micrograd: dynamic(() => import('../../sample/micrograd/main')),
-  complexCRT: dynamic(() => import('../../sample/complexCRT/main')),
+  complexCRT: dynamic(() => import('../../sample/fragment/complexCRT/main')),
   sdfCircle: dynamic(() => import('../../sample/sdfCircle/main')),
-  bitonicSort: dynamic(() => import('../../sample/bitonicSort/main')),
+  bitonicSort: dynamic(() => import('../../sample/compute/bitonicSort/main')),
 };
 
 export const fragmentPages = {
-  mixExample: dynamic(() => import('../../sample/mixExample/main')),
-  CRT: dynamic(() => import('../../sample/crt/main')),
-  grid: dynamic(() => import('../../sample/grid/main')),
-  cyberpunkGrid: dynamic(() => import('../../sample/cyberpunkGrid/main')),
-  complexCRT: dynamic(() => import('../../sample/complexCRT/main')),
+  mixExample: dynamic(() => import('../../sample/fragment/mixExample/main')),
+  CRT: dynamic(() => import('../../sample/fragment/crt/main')),
+  grid: dynamic(() => import('../../sample/fragment/grid/main')),
+  cyberpunkGrid: dynamic(
+    () => import('../../sample/fragment/cyberpunkGrid/main')
+  ),
+  complexCRT: dynamic(() => import('../../sample/fragment/complexCRT/main')),
 };
 
 export const vertexPages = {
@@ -40,8 +44,8 @@ export const vertexPages = {
 };
 
 export const computePages = {
-  computeBalls: dynamic(() => import('../../sample/computeBalls/main')),
-  bitonicSort: dynamic(() => import('../../sample/bitonicSort/main')),
+  computeBalls: dynamic(() => import('../../sample/compute/computeBalls/main')),
+  bitonicSort: dynamic(() => import('../../sample/compute/bitonicSort/main')),
 };
 
 function Page({ slug }: Props): JSX.Element {
