+++ src/sample/fragment/complexCRT/complexCRT.ts
@@ -1,46 +1,53 @@
-import {
-  BindGroupDescriptor,
-  createBindGroupDescriptor,
-} from '../../utils/bindGroup';
-import { Base2DRendererClass } from '../../utils/program/renderProgram';
-import { ShaderKeyInterface } from '../../utils/shaderUtils';
-import { BitonicDisplayShader, argKeys } from './renderShader';
+import { createBindGroupDescriptor } from '../../../utils/bindGroup';
+import { Base2DRendererClass } from '../../../utils/program/renderProgram';
+import { ComplexCRTShader, argKeys } from './shader';
+import { ShaderKeyInterface } from '../../../utils/shaderUtils';
 
-type BitonicDisplayRenderArgs = ShaderKeyInterface<typeof argKeys>;
+type ComplexCRTRendererArgs = ShaderKeyInterface<typeof argKeys> & {
+  textureName: string;
+};
 
-export default class BitonicDisplayRenderer extends Base2DRendererClass {
+export default class ComplexCRTRenderer extends Base2DRendererClass {
   static sourceInfo = {
     name: __filename.substring(__dirname.length + 1),
     contents: __SOURCE__,
   };
 
   switchBindGroup: (name: string) => void;
-  setArguments: (args: BitonicDisplayRenderArgs) => void;
-  computeBGDescript: BindGroupDescriptor;
+  setArguments: (args: Exclude<ComplexCRTRendererArgs, 'textureName'>) => void;
 
   constructor(
     device: GPUDevice,
     presentationFormat: GPUTextureFormat,
     renderPassDescriptor: GPURenderPassDescriptor,
     bindGroupNames: string[],
-    computeBGDescript: BindGroupDescriptor,
-    label: string
+    textures: GPUTexture[],
+    label: string,
+    debug = false
   ) {
     super();
     this.renderPassDescriptor = renderPassDescriptor;
-    this.computeBGDescript = computeBGDescript;
 
     const uniformBuffer = device.createBuffer({
       size: Float32Array.BYTES_PER_ELEMENT * argKeys.length,
       usage: GPUBufferUsage.UNIFORM | GPUBufferUsage.COPY_DST,
     });
 
+    const sampler = device.createSampler({
+      minFilter: 'linear',
+      magFilter: 'linear',
+    });
+
+    const resourceArr = textures.map((texture) => {
+      return [{ buffer: uniformBuffer }, sampler, texture.createView()];
+    });
+
     const bgDescript = createBindGroupDescriptor(
-      [0],
+      [0, 1, 2],
       [GPUShaderStage.FRAGMENT],
-      ['buffer'],
-      [{ type: 'uniform' }],
-      [[{ buffer: uniformBuffer }]],
+      ['buffer', 'sampler', 'texture'],
+      [{ type: 'uniform' }, { type: 'filtering' }, { sampleType: 'float' }],
+      resourceArr,
       label,
       device
     );
@@ -57,9 +64,9 @@ export default class BitonicDisplayRenderer extends Base2DRendererClass {
     this.pipeline = super.create2DRenderPipeline(
       device,
       label,
-      [bgDescript.bindGroupLayout, this.computeBGDescript.bindGroupLayout],
-      'WEBGL',
-      BitonicDisplayShader(),
+      [bgDescript.bindGroupLayout],
+      'NDCFlipped',
+      ComplexCRTShader(debug),
       presentationFormat
     );
 
@@ -68,16 +75,18 @@ export default class BitonicDisplayRenderer extends Base2DRendererClass {
       this.currentBindGroupName = name;
     };
 
-    this.setArguments = (args: BitonicDisplayRenderArgs) => {
+    this.setArguments = (args: ComplexCRTRendererArgs) => {
       super.setUniformArguments(device, uniformBuffer, args, argKeys);
     };
   }
 
-  startRun(commandEncoder: GPUCommandEncoder, args: BitonicDisplayRenderArgs) {
+  startRun(commandEncoder: GPUCommandEncoder, args: ComplexCRTRendererArgs) {
     this.setArguments(args);
+    if (args.textureName !== this.currentBindGroupName) {
+      this.switchBindGroup(args.textureName);
+    }
     super.executeRun(commandEncoder, this.renderPassDescriptor, this.pipeline, [
       this.currentBindGroup,
-      this.computeBGDescript.bindGroups[0],
     ]);
   }
 }
