+++ src/sample/fragment/cyberpunkGrid/cyberpunkGrid.ts
@@ -1,30 +1,26 @@
-import { createBindGroupDescriptor } from '../../utils/bindGroup';
-import crtFragWGSL from './crt.frag.wgsl';
-import crtDebugFragWGSL from './crtDebug.frag.wgsl';
-import { Base2DRendererClass } from '../../utils/program/renderProgram';
-import { ShaderKeyInterface } from '../../utils/shaderUtils';
+import { createBindGroupDescriptor } from '../../../utils/bindGroup';
+import { Base2DRendererClass } from '../../../utils/program/renderProgram';
+import CyberpunkCommonsWGSL from './cyberpunk_commons.wgsl';
+import { CyberpunkGridShader } from './shader';
+import { argKeys } from './shader';
+import { ShaderKeyInterface } from '../../../utils/shaderUtils';
 
-const argKeys = ['time', 'debugStep'];
+type CyberpunkGridRenderArgs = ShaderKeyInterface<typeof argKeys>;
 
-type CRTRendererArgs = ShaderKeyInterface<typeof argKeys> & {
-  textureName: string;
-};
-
-export default class CRTRenderer extends Base2DRendererClass {
+export default class CyberpunkGridRenderer extends Base2DRendererClass {
   static sourceInfo = {
     name: __filename.substring(__dirname.length + 1),
     contents: __SOURCE__,
   };
 
   switchBindGroup: (name: string) => void;
-  changeArgs: (args: CRTRendererArgs) => void;
+  changeArgs: (args: CyberpunkGridRenderArgs) => void;
 
   constructor(
     device: GPUDevice,
     presentationFormat: GPUTextureFormat,
     renderPassDescriptor: GPURenderPassDescriptor,
     bindGroupNames: string[],
-    textures: GPUTexture[],
     label: string,
     debug = false
   ) {
@@ -36,20 +32,13 @@ export default class CRTRenderer extends Base2DRendererClass {
       usage: GPUBufferUsage.UNIFORM | GPUBufferUsage.COPY_DST,
     });
 
-    const sampler = device.createSampler({
-      minFilter: 'linear',
-      magFilter: 'linear',
-    });
-
-    const resourceArr = textures.map((texture) => {
-      return [{ buffer: uniformBuffer }, sampler, texture.createView()];
-    });
+    const resourceArr = [[{ buffer: uniformBuffer }]];
 
     const bgDescript = createBindGroupDescriptor(
-      [0, 1, 2],
+      [0],
       [GPUShaderStage.FRAGMENT],
-      ['buffer', 'sampler', 'texture'],
-      [{ type: 'uniform' }, { type: 'filtering' }, { sampleType: 'float' }],
+      ['buffer'],
+      [{ type: 'uniform' }],
       resourceArr,
       label,
       device
@@ -68,12 +57,12 @@ export default class CRTRenderer extends Base2DRendererClass {
       device,
       label,
       [bgDescript.bindGroupLayout],
-      'WEBGPU',
-      debug ? crtDebugFragWGSL : crtFragWGSL,
+      'WEBGL',
+      CyberpunkGridShader(debug) + CyberpunkCommonsWGSL,
       presentationFormat
     );
 
-    this.changeArgs = (args: CRTRendererArgs) => {
+    this.changeArgs = (args: CyberpunkGridRenderArgs) => {
       super.setUniformArguments(device, uniformBuffer, args, argKeys);
     };
 
@@ -83,11 +72,8 @@ export default class CRTRenderer extends Base2DRendererClass {
     };
   }
 
-  startRun(commandEncoder: GPUCommandEncoder, args: CRTRendererArgs) {
+  startRun(commandEncoder: GPUCommandEncoder, args: CyberpunkGridRenderArgs) {
     this.changeArgs(args);
-    if (args.textureName !== this.currentBindGroupName) {
-      this.switchBindGroup(args.textureName);
-    }
     super.executeRun(commandEncoder, this.renderPassDescriptor, this.pipeline, [
       this.currentBindGroup,
     ]);
