+++ src/sample/fragment/grid/grid.frag.wgsl
