+++ src/sample/fragment/grid/grid.ts
@@ -1,6 +1,9 @@
-import fullscreenVertWGSL from '../../shaders/fullscreenWebGL.vert.wgsl';
-import { createBindGroupDescriptor } from '../../utils/bindGroup';
-import { Base2DRendererClass, BaseRenderer } from '../../utils/program/renderProgram';
+import fullscreenVertWGSL from '../../../shaders/fullscreenWebGL.vert.wgsl';
+import { createBindGroupDescriptor } from '../../../utils/bindGroup';
+import {
+  Base2DRendererClass,
+  BaseRenderer,
+} from '../../../utils/program/renderProgram';
 import gridFragWGSL from './grid.frag.wgsl';
 import gridDebugFragWGSL from './gridDebug.frag.wgsl';
