+++ src/sample/fragment/mixExample/mix.frag.wgsl
