+++ src/sample/micrograd/makemoreone.ts
@@ -14,8 +14,6 @@ export const createProbabilityArray = (arr: number[]) => {
   const probs = tf.tensor1d(tempProbs);
 
   const s = tf.multinomial(probs, 1, 2147483647);
-  console.log(arr);
-  console.log(probs);
   s.array().then((array) => {
     const item = array[0];
     console.log(item);
